app/requirements.txt
# --- Core FastAPI ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai==0.28.1
pydantic==2.7.1
pydantic-settings==2.2.1
grpcio==1.74.0
grpcio-tools==1.74.0
pact-python==2.3.3
pytest==8.3.3

# --- Database ---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Security ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1

# --- PDF Generation (НОВОЕ) ---
weasyprint==66.0

# --- Payments ---
stripe==8.3.0
